+++ packages/common/src/utils/getNestedProperty.ts
@@ -22,30 +22,3 @@ Original Code is the Infinite Reality Engine team.
 All portions of the code written by the Infinite Reality Engine team are Copyright © 2021-2023 
 Infinite Reality Engine. All Rights Reserved.
 */
-
-export function getNestedObject(object: any, propertyName: string) {
-  const props = propertyName.split('.')
-  let result = object
-
-  for (let i = 0; i < props.length - 1; i++) {
-    let isNumber = false
-
-    try {
-      Number(props[0])
-      isNumber = true
-    } catch (e) {
-      isNumber = false
-    }
-
-    let val = props[i] as string | number
-
-    if (isNumber) {
-      val = Number(val)
-    }
-
-    if (typeof result[props[i]] === 'undefined') result[props[i]] = {}
-    result = result[props[i]]
-  }
-
-  return { result, finalProp: props[props.length - 1] }
-}
