+++ packages/ecs/src/Timer.ts
@@ -23,7 +23,7 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { isClient } from '@ir-engine/common/src/utils/getEnvironment'
+import { isClient } from "@ir-engine/hyperflux"
 
 const performance: Performance = isClient ? window.performance : require('perf_hooks').performance
