+++ packages/server-core/src/core.test.ts
@@ -27,7 +27,7 @@ import assert from 'assert'
 
 import { destroyEngine } from '@ir-engine/ecs/src/Engine'
 
-import { createFeathersKoaApp } from './createApp'
+import { createFeathersKoaApp, tearDownAPI } from './createApp'
 
 describe('Core', () => {
   it('should initialise app', async () => {
@@ -35,7 +35,9 @@ describe('Core', () => {
     await app.setup()
     assert.ok(app.isSetup)
   })
-  after(() => {
-    return destroyEngine()
+  after(async () => {
+    await tearDownAPI()
+    destroyEngine()
   })
+
 })
