+++ packages/server-core/src/projects/project-check-source-destination-match/project-check-source-destination-match.test.ts
@@ -37,7 +37,7 @@ import { UserName, userPath } from '@ir-engine/common/src/schemas/user/user.sche
 import { destroyEngine } from '@ir-engine/ecs/src/Engine'
 
 import { Application, HookContext } from '../../../declarations'
-import { createFeathersKoaApp } from '../../createApp'
+import { createFeathersKoaApp, tearDownAPI } from '../../createApp'
 import { identityProviderDataResolver } from '../../user/identity-provider/identity-provider.resolvers'
 import { getRepoManifestJson1, getRepoManifestJson2 } from '../../util/mockOctokitResponses'
 
@@ -94,7 +94,10 @@ describe('project-check-source-destination-match.test', () => {
     )
   })
 
-  after(() => destroyEngine())
+  after(async () => {
+    await tearDownAPI()
+    destroyEngine()
+  })
 
   it('should match source and destination contents with same repos', async () => {
     nock('https://api.github.com')
