+++ packages/server-core/src/projects/project-check-unfetched-commit/project-check-unfetched-commit.test.ts
@@ -36,7 +36,7 @@ import { UserName, userPath } from '@ir-engine/common/src/schemas/user/user.sche
 import { destroyEngine } from '@ir-engine/ecs/src/Engine'
 
 import { Application, HookContext } from '../../../declarations'
-import { createFeathersKoaApp } from '../../createApp'
+import { createFeathersKoaApp, tearDownAPI } from '../../createApp'
 import { identityProviderDataResolver } from '../../user/identity-provider/identity-provider.resolvers'
 import { getRepoManifestJson1, getTestRepoCommit } from '../../util/mockOctokitResponses'
 
@@ -54,9 +54,7 @@ describe('project-check-unfetched-commit.test', () => {
   before(async () => {
     app = createFeathersKoaApp()
     await app.setup()
-  })
-
-  before(async () => {
+    
     const name = ('test-project-check-unfetched-commit-user-name-' + uuidv4()) as UserName
 
     const avatar = await app.service(avatarPath).create({
@@ -85,7 +83,10 @@ describe('project-check-unfetched-commit.test', () => {
     )
   })
 
-  after(() => destroyEngine())
+  after(async () => {
+    await tearDownAPI()
+    destroyEngine()
+  })
 
   it('should get the commit data', async () => {
     nock('https://api.github.com')
