+++ packages/server-core/src/social/location/location.test.ts
@@ -32,7 +32,7 @@ import { destroyEngine } from '@ir-engine/ecs/src/Engine'
 
 import { staticResourcePath } from '@ir-engine/common/src/schema.type.module'
 import { Application } from '../../../declarations'
-import { createFeathersKoaApp } from '../../createApp'
+import { createFeathersKoaApp, tearDownAPI } from '../../createApp'
 import { LocationParams } from './location.class'
 
 const params = { isInternal: true } as LocationParams
@@ -46,8 +46,9 @@ describe('location.test', () => {
     await app.setup()
   })
 
-  after(() => {
-    return destroyEngine()
+  after(async () => {
+    await tearDownAPI()
+    destroyEngine()
   })
 
   it('should create a new location', async () => {
